Add debug flag, cleanup special keys, separate codes
Arrow keys still don't work properly. That will require
keeping track of the cursors position on the line. Normally
it would be length of the current line, but will vary
if a use presses one of the arrows.

Backspace is handled properly now but DEL is not. It generates
a special escape code (which somewhat varies on terminal emulator
but in general is consistantly ^[3~) and needs to be handled
with other escape codes. Also it kind of requires arrow keys
to work properly otherwise a user can't go over a letter to
delete.

The flag -hexdebug will print the Hex codes of pressed keys.
As the name suggests it's strictly for debugging purposes.