Started Lisp VM, defvar, print
Refactored ast to objects, make everything references
for GC. Implemented the beginnings of a VM. It's a tree
walking interpreter. It'll be good enough for now. Maybe
later I can rewrite it as a bytecode VM.

The VM has a global symbol table, no local tables but the Env
struct can accept a parent.

The only implemented functions are defvar and print. User functions
are not available yet.