Implemented free form commands and "call"
The "call" function invokes a system command with
given arguments. It can be used explicitly as a
function or implicitly when calling a system
command that doesn't exist as a function in Lazuli.

Example:

(call ls -lah)
(ls -lah)

The above two commands will do the same thing so long
as there's no defined function named "ls" in the current
scope. Functions defined in the shell vm override external
commands.