Expand ~, implement pipe, fix cd
When changing directories, tilde will now expand
to the user's home directory. ~user is not supported.

The cd command no longer prints an empty list.

Commands can now be piped together using the pipe function.
Pipe takes lists of commands. The arguments do not need
to be quoted as pipe doesn't eval it arguments. It does
eval the members of each list so symbols and such
can be expanded.